// typechain-types/DayGlimpse.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type {
  BaseContract,
  BigNumberish,
  BytesLike,
  FunctionFragment,
  Result,
  Interface,
  AddressLike,
  ContractRunner,
  ContractMethod,
  Listener,
} from "ethers";
import type {
  TypedContractEvent,
  TypedDeferredTopicFilter,
  TypedEventLog,
  TypedListener,
  TypedContractMethod,
} from "./common";

export interface DayGlimpseInterface extends Interface {
  getFunction(
    nameOrSignature: "getDayGlimpse" | "setDayGlimpse"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "getDayGlimpse",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "setDayGlimpse",
    values: [BytesLike, BigNumberish, boolean]
  ): string;

  decodeFunctionResult(
    functionFragment: "getDayGlimpse",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setDayGlimpse",
    data: BytesLike
  ): Result;
}

export interface DayGlimpse extends BaseContract {
  connect(runner?: ContractRunner | null): DayGlimpse;
  waitForDeployment(): Promise<this>;

  interface: DayGlimpseInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getDayGlimpse: TypedContractMethod<
    [_profile: AddressLike],
    [
      [string, bigint, boolean, boolean] & {
        storageHash: string;
        timestamp: bigint;
        isPrivate: boolean;
        isActive: boolean;
      }
    ],
    "view"
  >;

  setDayGlimpse: TypedContractMethod<
    [_storageHash: BytesLike, _timestamp: BigNumberish, _isPrivate: boolean],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getDayGlimpse"
  ): TypedContractMethod<
    [_profile: AddressLike],
    [
      [string, bigint, boolean, boolean] & {
        storageHash: string;
        timestamp: bigint;
        isPrivate: boolean;
        isActive: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "setDayGlimpse"
  ): TypedContractMethod<
    [_storageHash: BytesLike, _timestamp: BigNumberish, _isPrivate: boolean],
    [void],
    "nonpayable"
  >;

  filters: {};
}